moved plugin decompilation out of MainWindowCommands
