+++ UI/MainWindow/MainWindowCommands.cs
@@ -12,16 +12,12 @@ using SPCode.UI.Components;
 using SPCode.UI.Windows;
 using SPCode.Utils;
 using SPCode.Utils.SPSyntaxTidy;
-using static SPCode.Utils.JavaUtils;
+using static SPCode.Utils.JavaInstallation;
 
 namespace SPCode.UI
 {
     public partial class MainWindow
     {
-
-        private ProgressDialogController dwJavaInCourse;
-        private readonly string OutFile = Environment.ExpandEnvironmentVariables(Constants.JavaDownloadFile);
-        private readonly string JavaLink = Environment.Is64BitOperatingSystem ? Constants.JavaDownloadSite64 : Constants.JavaDownloadSite32;
 
         public EditorElement GetCurrentEditorElement()
         {
@@ -320,23 +316,6 @@ namespace SPCode.UI
                         numOfSpacesOrTabsBefore = ee.editor.Document.GetText(line).Count(c => c == ' ' || c == '\t');
                     }
 
-#if DEBUG
-                    Debug.WriteLine($"Curser offset before format: {currentCaret}");
-
-                    // Where is out curser?
-                    if (currentCaret == line.Offset)
-                    {
-                        Debug.WriteLine("Curser is at the start of the line");
-                    }
-                    else if (currentCaret == line.EndOffset)
-                    {
-                        Debug.WriteLine("Curser is at the end of the line");
-                    }
-                    else
-                    {
-                        Debug.WriteLine("Curser is somewhere in the middle of the line");
-                    }
-#endif
                     // Formatting Start //
                     ee.editor.Document.BeginUpdate();
                     var source = ee.editor.Text;
@@ -354,9 +333,6 @@ namespace SPCode.UI
                     {
                         var numOfSpacesOrTabsAfter = ee.editor.Document.GetText(line).Count(c => c == ' ' || c == '\t');
                         newCaretPos += curserLinePos + (numOfSpacesOrTabsAfter - numOfSpacesOrTabsBefore);
-#if DEBUG
-                        Debug.WriteLine($"Curser offset after format: {newCaretPos}");
-#endif
                     }
                     ee.editor.TextArea.Caret.Offset = newCaretPos;
                 }
@@ -365,174 +341,9 @@ namespace SPCode.UI
 
         private async void Command_Decompile(MainWindow win)
         {
-            // First we check the java version of the user, and act accordingly
-
-            ProgressDialogController checkingJavaDialog = null;
-            if (win != null)
-            {
-                checkingJavaDialog = await this.ShowProgressAsync(Program.Translations.GetLanguage("JavaInstallCheck") + "...",
-                    "", false, MetroDialogOptions);
-                ProcessUITasks();
-            }
-            var ju = new JavaUtils();
-            switch (ju.GetJavaStatus())
-            {
-                case JavaResults.Absent:
-                    {
-                        // If java is not installed, offer to download it
-                        await checkingJavaDialog.CloseAsync();
-                        if (await this.ShowMessageAsync(Program.Translations.GetLanguage("JavaNotFoundTitle"),
-                            Program.Translations.GetLanguage("JavaNotFoundMessage"),
-                            MessageDialogStyle.AffirmativeAndNegative, MetroDialogOptions) == MessageDialogResult.Affirmative)
-                        {
-                            await InstallJava();
-                        }
-                        return;
-                    }
-                case JavaResults.Outdated:
-                    {
-                        // If java is outdated, offer to upgrade it
-                        await checkingJavaDialog.CloseAsync();
-                        if (await this.ShowMessageAsync(Program.Translations.GetLanguage("JavaOutdatedTitle"),
-                             Program.Translations.GetLanguage("JavaOutdatedMessage"),
-                             MessageDialogStyle.AffirmativeAndNegative, MetroDialogOptions) == MessageDialogResult.Affirmative)
-                        {
-                            await InstallJava();
-                        }
-                        return;
-                    }
-                case JavaResults.Correct:
-                    {
-                        // Move on
-                        await checkingJavaDialog.CloseAsync();
-                        break;
-                    }
-            }
-
-            // Pick file for decompiling
-            var ofd = new OpenFileDialog
-            {
-                Filter = "Sourcepawn Plugins (*.smx)|*.smx",
-                Title = Program.Translations.GetLanguage("ChDecomp")
-            };
-            var result = ofd.ShowDialog();
-
-            if (result.Value && !string.IsNullOrWhiteSpace(ofd.FileName))
-            {
-                var fInfo = new FileInfo(ofd.FileName);
-                if (fInfo.Exists)
-                {
-                    ProgressDialogController task = null;
-                    if (win != null)
-                    {
-                        task = await this.ShowProgressAsync(Program.Translations.GetLanguage("Decompiling") + "...",
-                            fInfo.FullName, false, MetroDialogOptions);
-                        ProcessUITasks();
-                    }
-
-                    // Prepare Lysis execution
-                    var destFile = fInfo.FullName + ".sp";
-                    var standardOutput = new StringBuilder();
-                    using var process = new Process();
-                    process.StartInfo.WorkingDirectory = Paths.GetLysisDirectory();
-                    process.StartInfo.UseShellExecute = false;
-                    process.StartInfo.RedirectStandardOutput = true;
-                    process.StartInfo.CreateNoWindow = true;
-                    process.StartInfo.FileName = "java";
-
-                    process.StartInfo.Arguments = $"-jar lysis-java.jar \"{fInfo.FullName}\"";
-
-                    // Execute Lysis, read and store output
-                    try
-                    {
-                        process.Start();
-                        while (!process.HasExited)
-                        {
-                            standardOutput.Append(process.StandardOutput.ReadToEnd());
-                        }
-                        standardOutput.Append(process.StandardOutput.ReadToEnd());
-                        File.WriteAllText(destFile, standardOutput.ToString(), Encoding.UTF8);
-                    }
-                    catch (Exception ex)
-                    {
-                        await this.ShowMessageAsync($"{fInfo.Name} {Program.Translations.GetLanguage("FailedToDecompile")}",
-                            $"{ex.Message}", MessageDialogStyle.Affirmative,
-                        MetroDialogOptions);
-                    }
-
-                    // Load the decompiled file to SPCode
-                    TryLoadSourceFile(destFile, true, false, true);
-                    if (task != null)
-                    {
-                        await task.CloseAsync();
-                    }
-                }
-            }
+            var decomp = new DecompileUtil(win, MetroDialogOptions);
+            await decomp.DecompilePlugin();
         }
-
-        #region Java Installation
-
-        private async System.Threading.Tasks.Task InstallJava()
-        {
-            // Spawn progress dialog when downloading Java
-            dwJavaInCourse = await this.ShowProgressAsync(Program.Translations.GetLanguage("DownloadingJava") + "...",
-                Program.Translations.GetLanguage("FetchingJava"), false, MetroDialogOptions);
-            dwJavaInCourse.SetProgress(0.0);
-            ProcessUITasks();
-
-            // Setting up event callbacks to change download percentage, amount downloaded and amount left
-            using var wc = new WebClient();
-            wc.DownloadProgressChanged += DownloadProgressed;
-            wc.DownloadFileCompleted += DownloadCompleted;
-            wc.DownloadFileAsync(new Uri(JavaLink), OutFile);
-        }
-
-        private void DownloadProgressed(object sender, DownloadProgressChangedEventArgs e)
-        {
-            // Handles percentage and MB downloaded/left
-            dwJavaInCourse.SetMessage(
-                $"{e.ProgressPercentage}% {Program.Translations.GetLanguage("AmountCompleted")}, " +
-                $"{Program.Translations.GetLanguage("AmountDownloaded")} {Math.Round(ByteSize.FromBytes(e.BytesReceived).MegaBytes),0} MB / " +
-                $"{Math.Round(ByteSize.FromBytes(e.TotalBytesToReceive).MegaBytes),0} MB");
-
-            // Handles progress bar
-            dwJavaInCourse.SetProgress(e.ProgressPercentage * 0.01d);
-        }
-
-        private async void DownloadCompleted(object sender, AsyncCompletedEventArgs e)
-        {
-            await dwJavaInCourse.CloseAsync();
-            if (File.Exists(OutFile))
-            {
-                // If file downloaded properly, it should open
-                Process.Start(OutFile);
-                await this.ShowMessageAsync(
-                    Program.Translations.GetLanguage("JavaOpened"),
-                    Program.Translations.GetLanguage("JavaSuggestRestart"),
-                    MessageDialogStyle.Affirmative);
-            }
-            else
-            {
-                // Otherwise, just offer a manual download
-                if (await this.ShowMessageAsync(
-                    Program.Translations.GetLanguage("JavaDownErrorTitle"),
-                    Program.Translations.GetLanguage("JavaDownErrorMessage"),
-                    MessageDialogStyle.AffirmativeAndNegative, MetroDialogOptions) == MessageDialogResult.Affirmative)
-                {
-                    Process.Start(new ProcessStartInfo
-                    {
-                        FileName = JavaLink,
-                        UseShellExecute = true
-                    });
-                    await this.ShowMessageAsync(
-                    Program.Translations.GetLanguage("JavaOpenedBrowser"),
-                    Program.Translations.GetLanguage("JavaSuggestRestart"),
-                    MessageDialogStyle.Affirmative);
-                }
-            }
-        }
-
-        #endregion
 
         private void Command_OpenSPDef()
         {
